+++ examples/unicode.rs
@@ -8,9 +8,9 @@ const CONSOLE_HEIGHT: u32 = 25;
 struct MyRoguelike {}
 
 impl Engine for MyRoguelike {
-    fn init(&mut self, _api: &mut DoryenApi) {}
-    fn update(&mut self, _api: &mut DoryenApi) {}
-    fn render(&mut self, api: &mut DoryenApi) {
+    fn init(&mut self, _api: &mut dyn DoryenApi) {}
+    fn update(&mut self, _api: &mut dyn DoryenApi) {}
+    fn render(&mut self, api: &mut dyn DoryenApi) {
         let con = api.con();
         con.clear(Some((32, 16, 0, 255)), Some((255, 240, 224, 255)), None);
         con.ascii(
@@ -24,7 +24,7 @@ impl Engine for MyRoguelike {
         con.print(25, 5, "Φοβερός!", TextAlign::Center, None, None);
         con.print(8, 12, "ça, c'est énorme!", TextAlign::Left, None, None);
     }
-    fn resize(&mut self, _api: &mut DoryenApi) {}
+    fn resize(&mut self, _api: &mut dyn DoryenApi) {}
 }
 
 fn main() {